Restructure site: Decentralize 'pages' folder.
Now the 'pages' folder only holds toplevel pages of the website. All
other folders and subfolders are in the root of the repository.

This was mainly done to fix the problem of static files not copying from
the pages/* folders to _site.

Also separated glyph conventions (docs) as different pages and updated
pages accordingly.

All other functionality remains the same.