[themes] Create array of themes.
Themes are now an array in _includes/head.html. If theme name in YAML
frontmatter exists in the list, it is used as the theme for that page.

Method to add more themes:

1. Create CSS files in css/ with name freetype-{theme-name}.css

2. If the theme has a completely new color, copy freetype-blue.css to
the file and make changes to the required colors.

3. If the theme is only a shade of an existing color theme, import the
original color file, and override with custom rules. Refer to
'freetype-dark-green.css' for an idea of how this works.

4. Add the theme name (without 'freetype-' part) to the array in
_includes/head.html. Done!

* _includes/footer.html: Change date format.

* _includes/head*.html: Change CSS references.